Use same query string parsing for 'RouterRequest' and 'BodyParser'
